feat: Implementação de cache em todos os serviços: Notification, Payment e Reconciliations
- Adicionado cache para otimizar consultas e reduzir carga no banco de dados.
- Implementado mecanismo de cache nos serviços de Notification, Payment e Reconciliations.
- Utilização de [Redis, MemoryCache].
- Configuração do tempo de expiração do cache conforme necessidade de cada serviço.
- Melhorias na performance e escalabilidade dos serviços.

+++ FinancePlatform.API/Application/Services/NotificationService.cs
@@ -8,7 +8,6 @@ using FinancePlatform.API.Presentation.DTOs.ViewModel;
 using FluentValidation;
 using Mapster;
 using MapsterMapper;
-using System.Security.Principal;
 
 namespace FinancePlatform.API.Application.Services
 {
@@ -16,6 +15,7 @@ namespace FinancePlatform.API.Application.Services
     {
         private readonly INotificationRepository _notificationRepository;
         private readonly IValidator<Notification> _validator;
+        private readonly IValidator<Guid> _guidValidator;
         private readonly IEntityUpdateStrategy _entityUpdateStrategy;
         private readonly IMapper _mapper;
         private readonly CacheService _cacheService;
@@ -23,28 +23,63 @@ namespace FinancePlatform.API.Application.Services
         public NotificationService(INotificationRepository notificationRepository, 
                                    IEntityUpdateStrategy entityUpdateStrategy,
                                    IValidator<Notification> validator,
+                                   IValidator<Guid> guidValidator,
                                    IMapper mapper,
                                    CacheService cacheService)
         {
             _notificationRepository = notificationRepository;
             _entityUpdateStrategy = entityUpdateStrategy;
+            _guidValidator = guidValidator;
             _validator = validator;
             _mapper = mapper;
             _cacheService = cacheService;
         }
 
-        public async Task<NotificationViewModel?> GetNotificationByIdAsync(Guid idNotification)
+        public async Task<NotificationViewModel?> FindNotificationByIdAsync(Guid notificationId)
         {
-            var notification = await _notificationRepository.FindByIdAsync(idNotification);
+            var validationResult = _guidValidator.Validate(notificationId);
 
-            return _mapper.Map<NotificationViewModel>(notification);
+            if (!validationResult.IsValid) return null;
+
+            string notificationCacheKey = $"account:{notificationId}";
+
+            var cachedNotification = await _cacheService.GetAsync<Account>(notificationCacheKey);
+            
+            if (cachedNotification != null)
+            {
+                return _mapper.Map<NotificationViewModel>(cachedNotification);
+            }
+
+            var existingNotification = await _notificationRepository.FindByIdAsync(notificationId);
+
+            if (existingNotification == null) return null;
+
+            await _cacheService.SetAsync(notificationCacheKey, existingNotification);
+
+            return _mapper.Map<NotificationViewModel>(existingNotification);
         }
 
-        public async Task<List<NotificationViewModel>?> GetAllNotificationsAsync()
+        public async Task<List<NotificationViewModel>?> FindAllNotificationsAsync()
         {
-            var notifications = await _notificationRepository.FindAllAsync();
+            string notificationListCacheKey = "notifications:list";
 
-            return _mapper.Map<List<NotificationViewModel>>(notifications);
+            var cachedNotifications = await _cacheService.GetAsync<List<Account>>(notificationListCacheKey);
+            
+            if (cachedNotifications != null)
+            {
+                return _mapper.Map<List<NotificationViewModel>>(cachedNotifications);
+            }
+
+            var existingNotifications = await _notificationRepository.FindAllAsync();
+            
+            if (existingNotifications == null || existingNotifications.Count == 0)
+            {
+                return null;
+            }
+
+            await _cacheService.SetAsync(notificationListCacheKey, existingNotifications);
+
+            return _mapper.Map<List<NotificationViewModel>>(existingNotifications);
         }
 
         public async Task<Notification?> CreateNotificationAsync(NotificationInputModel model)
@@ -54,28 +89,54 @@ namespace FinancePlatform.API.Application.Services
 
             if (!validator.IsValid) return null;
 
-            await _notificationRepository.Add(notification);
+            var createdNotification = await _notificationRepository.Add(notification);
+
+            if (createdNotification != null)
+            {
+                string accountCacheKey = $"notification:{createdNotification.Id}";
+
+                await _cacheService.SetAsync(accountCacheKey, createdNotification);
+            }
             return notification;
         }
 
         public async Task<Notification?> UpdateAsync(Guid notificationId, Dictionary<string, object> updateRequest)
         {
-            Notification notification = await _notificationRepository.FindByIdAsync(notificationId);
+            var validationResult = _guidValidator.Validate(notificationId);
+
+            if (!validationResult.IsValid) return null;
+
+            var notification = await _notificationRepository.FindByIdAsync(notificationId);
+            
             if (notification == null) return null;
 
-            _entityUpdateStrategy.UpdateEntityFields(notification, updateRequest);
+            var isUpdateSuccessful = _entityUpdateStrategy.UpdateEntityFields(notification, updateRequest);
 
-            await _notificationRepository.Update(notification);
-
+            if (isUpdateSuccessful)
+            {
+                await _notificationRepository.Update(notification);
+                
+                string accountCacheKey = $"notification:{notificationId}";
+                await _cacheService.UpdateCacheIfNeededAsync(accountCacheKey, notification);
+            }
             return notification;
         }
 
-        public async Task<bool> DeleteNotificationAsync(Guid id)
+        public async Task<bool> DeleteNotificationAsync(Guid notificationId)
         {
-            var notification = await _notificationRepository.FindByIdAsync(id);
+            var validationResult = _guidValidator.Validate(notificationId);
+
+            if (!validationResult.IsValid) return false;
+
+            var notification = await _notificationRepository.FindByIdAsync(notificationId);
+            
             if (notification == null) return false;
 
             _notificationRepository.Delete(notification);
+
+            string notificationCacheKey = $"notification:{notificationId}";
+            await _cacheService.RemoveFromCacheIfNeededAsync<Notification>(notificationCacheKey);
+
             return true;
         }
     }
