+++ FinancePlatform.API/Application/Services/ReconciliationService.cs
@@ -1,6 +1,7 @@
 ﻿using FinancePlatform.API.Application.Interfaces.Repositories;
 using FinancePlatform.API.Application.Interfaces.Services;
 using FinancePlatform.API.Application.Interfaces.Utils;
+using FinancePlatform.API.Application.Services.Cache;
 using FinancePlatform.API.Domain.Entities;
 using FinancePlatform.API.Presentation.DTOs.InputModel;
 using FinancePlatform.API.Presentation.DTOs.ViewModel;
@@ -14,18 +15,24 @@ namespace FinancePlatform.API.Application.Services
     {
         private readonly IReconciliationRepository _reconciliationRepository;
         private readonly IValidator<Reconciliation> _validator;
+        private readonly IValidator<Guid> _guidValidator;
         private readonly IEntityUpdateStrategy _entityUpdateStrategy;
         private readonly IMapper _mapper;
+        private readonly CacheService _cacheService;
 
         public ReconciliationService(IReconciliationRepository reconciliationRepository,
                                      IEntityUpdateStrategy entityUpdateStrategy,
+                                     IValidator<Guid> guidValidator,
                                      IValidator<Reconciliation> validator,
-                                     IMapper mapper)
+                                     IMapper mapper,
+                                     CacheService cacheService)
         {
             _reconciliationRepository = reconciliationRepository;
             _entityUpdateStrategy = entityUpdateStrategy;
+            _guidValidator = guidValidator;
             _validator = validator;
             _mapper = mapper;
+            _cacheService = cacheService;
         }
 
         public async Task<Reconciliation?> CreateReconciliation(ReconciliationInputModel model)
@@ -35,41 +42,99 @@ namespace FinancePlatform.API.Application.Services
 
             if (!validator.IsValid) return null;
 
-            return await _reconciliationRepository.AddAsync(reconciliation);
+            var createdReconciliation = await _reconciliationRepository.AddAsync(reconciliation);
+
+            if (createdReconciliation != null)
+            {
+                string reconciliationCacheKey = $"reconciliation:{createdReconciliation.Id}";
+
+                await _cacheService.SetAsync(reconciliationCacheKey, createdReconciliation);
+            }
+
+            return createdReconciliation;
         }
 
-        public async Task<ReconciliationViewModel?> GetReconciliationByIdAsync(Guid id)
+        public async Task<ReconciliationViewModel?> FindReconciliationByIdAsync(Guid reconciliationId)
         {
-            var reconciliation = await _reconciliationRepository.FindByIdAsync(id);
+            var validationResult = _guidValidator.Validate(reconciliationId);
 
-            return _mapper.Map<ReconciliationViewModel>(reconciliation);
+            if (!validationResult.IsValid) return null;
+
+            string reconciliationCacheKey = $"reconciliation:{reconciliationId}";
+
+            var cachedReconciliation = await _cacheService.GetAsync<Reconciliation>(reconciliationCacheKey);
+            if (cachedReconciliation != null)
+            {
+                return _mapper.Map<ReconciliationViewModel>(cachedReconciliation);
+            }
+
+            var existingReconciliation = await _reconciliationRepository.FindByIdAsync(reconciliationId);
+
+            if (existingReconciliation == null)
+            {
+                return null;
+            }
+            await _cacheService.SetAsync(reconciliationCacheKey, existingReconciliation);
+
+            return _mapper.Map<ReconciliationViewModel>(existingReconciliation);
         }
 
-        public async Task<List<ReconciliationViewModel>?> GetAllReconciliationsAsync()
+        public async Task<List<ReconciliationViewModel>?> FindAllReconciliationsAsync()
         {
-            var reconciliations = await _reconciliationRepository.FindAllAsync();
+            string reconciliationListCacheKey = "reconciliations:list";
 
-            return _mapper.Map<List<ReconciliationViewModel>>(reconciliations);
+            var cachedReconciliations = await _cacheService.GetAsync<List<Reconciliation>>(reconciliationListCacheKey);
+
+            if (cachedReconciliations != null)
+            {
+                return _mapper.Map<List<ReconciliationViewModel>>(cachedReconciliations);
+            }
+
+            var existingReconciliation = await _reconciliationRepository.FindAllAsync();
+            
+            if (existingReconciliation == null || existingReconciliation.Count == 0) return null;
+
+            await _cacheService.SetAsync(reconciliationListCacheKey, existingReconciliation);
+
+            return _mapper.Map<List<ReconciliationViewModel>>(existingReconciliation);
         }
 
         public async Task<Reconciliation?> UpdateReconciliationAsync(Guid reconciliationId, Dictionary<string, object> updateRequest)
         {
-            var reconciliationResult = await _reconciliationRepository.FindByIdAsync(reconciliationId);
-            if (reconciliationResult == null) return null;
+            var validationResult = _guidValidator.Validate(reconciliationId);
 
-            _entityUpdateStrategy.UpdateEntityFields(reconciliationResult, updateRequest);
+            if (!validationResult.IsValid) return null;
 
-            var reconciliationUpdate = await _reconciliationRepository.UpdateAsync(reconciliationResult);
+            var reconciliation = await _reconciliationRepository.FindByIdAsync(reconciliationId);
 
-            return reconciliationUpdate;
+            if (reconciliation == null) return null;
+
+            var isUpdateSuccessful = _entityUpdateStrategy.UpdateEntityFields(reconciliation, updateRequest);
+
+            if (isUpdateSuccessful)
+            {
+                await _reconciliationRepository.UpdateAsync(reconciliation);
+
+                string reconciliationCacheKey = $"reconciliation:{reconciliationId}";
+                await _cacheService.UpdateCacheIfNeededAsync(reconciliationCacheKey, reconciliation);
+            }
+            return reconciliation;
         }
 
         public async Task<bool> DeleteReconciliationAsync(Guid reconciliationId)
         {
+            var validationResult = _guidValidator.Validate(reconciliationId);
+
+            if (!validationResult.IsValid) return false;
+
             var reconciliation = await _reconciliationRepository.FindByIdAsync(reconciliationId);
             if (reconciliation == null) return false;
 
             _reconciliationRepository.Delete(reconciliation);
+
+            string accountCacheKey = $"reconciliation:{reconciliationId}";
+            await _cacheService.RemoveFromCacheIfNeededAsync<Account>(accountCacheKey);
+            
             return true;
         }
     }
